Add generate_comments_copy_function db function
This fills a narrow gap where a Public API wrapper function should
simply take on the comments of the wrapped private DB function.

Really, this function is probably a bad idea as it moves the in database
documentation away from the function where it might be used.  Still
duplication isn't great either and while having to go look up the docs
somewhere else is annoying, forgetting to update manually copied
comments when they change can be outright misleading.  So err on the
side of accepting disconnected comments for now.