Refactor access account associations in MSBMS BRM
Prior to this commit, Access Account associations from the MCP database
where assigned to a table in the MscmpSystInteractions Component, which
then depended on the MscmpBrmPerson and other BRM modules to link the
Access Account to the Person.

Ultimately this doesn't make sense.  If we mean it when we say that a
single person record in the MscmpBrmPerson Component should represent a
single person in the real world, we should also mean that their login
association is correctly and uniquely assigned to that person record.
Any question of management and Person record ownership simply follows
from the Person record itself: the Access Account merely identifies and
authenticates that the person is who they say they are.

So remove the references from the MscmpSystInteractions module (which
was an odd place to have it anyway) and correctly make it an association
with the MscmpBrmPerson Component.