Remove Feature Map references from Components
The "Feature Map" Component was removed some time ago, but there are
still some references out there.

This commit removes those references from MscmpSystDocnum and
MscmpSystFormats components, which I believe are the last references.

The Feature Map idea isn't gone, but the implementation is still an open
question.  I think it may just be a MscmpSystHierarchy implementation,
but it also might simply be MscmpSystNav data, too.  Ultimately we'll
cross that bridge when we get there at this point.