Refactor README.md / moduledoc in project APIs
This is about the Elixir Components incorporating the README.md files
into the moduledoc for the main API modules.

Previously, the comments were doing this incorporation, but it was
piecemeal component to component.

This comment ensures that we're doing it across the project and doing it
so that we aren't duplicating the title from the readme since ExDoc also
adds the title.