Refactor MscmpSystNav syst_nav_actions; add trigs
There are two significant changes in this commit:

  - Change `syst_action*` names to `syst_nav_action*`

  - Add API view triggers

The naming change arises because "action" is too generic when looking at
the database tables.  For example, MssubMcp loads the database tables of
many modules, but the tables are not explicitly named for their source
Component.  The term "action" is now used to define two similar/related,
but different, concepts coming from two different Components.  While the
naming collision could be avoided in an acceptable manner in this case,
it still could crop up again.  Since the work for more explicit naming
is already done (prior to the collision avoidance thinking being done)
let's keep the more explicit naming and if it becomes cumbersome we can
always change it back later.

Add the API view triggers for MscmpSystNav.  This wasn't done previously
and was done while addressing the naming changes.