Add MscmpSystInteraction database schema
Currently the Integration and Doctest buildplans add MscmpSystPerms
related unit test seed data.  I'm not sure this is really correct and so
it may change, but for now it helped get through validating the form of
the buildplans and that the MscmpSystInteraction integration and
doctests could be run in principle.