Add MscmpSystInteraction dependencies
This commit adds MscmpSystDb and MscmpSystPerms as dependencies of
MscmpSystInteraction.

The thinking here is to start early on the features which allow user
reassignment of data/interaction permissioning or permission definitions
for fully user defined functionality.  While the intent is not really
to offer this completely end to end in the application just yet, getting
some of the initial ideas in place should provide an easier path for
such support later.