Add MscmpSystInteraction State data type defs
One of the tasks of this Component is caching and controlling the
"Original Data " of the Interaction Context at runtime.  In order to
do that, we will have some sort of in-memory state (probably a
GenServer) to contain the actual Original Data as well as a little meta
data to help manage the runtime concerns relative to that state.

In this commit we add a struct which will hold the state and define a
type indicating whether the state data represents a new record or is
backed by persistent data.