Add MscmpSystEnums.TestSupport typespec; fix bug
This just adds a missing typespec that ElixirLS was (rightfully)
complaining about.

There was also a minor bugfix; the default name of the "Context
Registry" used an older form of the registry name which was inconsistent
with the intended default/common usage value.  In practice this is
unlikely to result in any runtime difference as, right now, there is no
code which expects the default name to be used... there's even an
argument that it should be removed.

Finally remove unneeded match on final function call.  The match doesn't
actually evaluate any condition and we get the right-hand side as the
return value anyway... so just let it return with no match.