Remove MscmpSystSettings.Runtime.DevSupport & API
I always hated having to have this for other modules to use.  During the
work to make contexts work with GenServer-like naming conventions, it
became apparent I could simply have this dev/test time available as
macros which could then be compiled into the modules needing the support
rather than have a full-fledged public API.  This means that, while the
macro definitions will continue to exist in releases, the API will
typically not be available since the places that need it typically are
only compiled during dev/test builds and can make use of them privately.

Also in this commit are changes to make this Component compatible with
similar dev support changes in MscmpSystDb.