Refactor MscmpSystNav for dev support changes
MscmpSystDb & MscmpSystEnums changed the way that development support
services are provided for client Components and this commit is chiefly
aimed at making MscmpSystNav compatible with those changes.

MscmpSystNav is still in early development and I can't test that the
changes are working without completing other, unrelated work on this
module.  Since the dev support changes are largely cut/paste from
elsewhere, we'll assume that they're correct for now and do detailed
validation later when it becomes needed for completing the Component.