Refactor mscmp_syst_db NimbleOptions  reorg
This is generally aimed at refactoring how NimbleOptions is implemented.

The primary idea is that all NimbleOptions option definition,
defaults, and validation should happen at the Public API layer rather
that directly in the implementation.  The issue, more pronounced in
other Components, is that we were doing a lot of duplication of option
definition and checking as higher level functions would have to pass
options to lower level functions, which could be called independently
and thus also require a params validation step.

Handling all of the options stuff at the public API layer means the
defaults, documentation, and definition of the hoops a Component user
would have to jump through is done where a Component user would look
to see what's right and wrong.  While it does mean that there's some
logic now at the API layer, that logic is genuinely focused on how the
user interacts with the Component and thus it feels appropriate to have
it here; it doesn't feel like a compromise or a "least bad" selection.

Update the NimbleOptions check, reformat some code and comments to
help with code navigation.  Bump dependencies and regenerate
documentation.