Refactor mscmp_syst_instance; concept reorg
For this component, the reorganization is larger than just bringing in
NimbleOptions, option definition and validation at the API boundary and
accommodation of the other recent refactoring of dependencies.  There is
a conceptual shift in how Applications and Instances are understood.

Prior to now, the idea of an Application and its Instances has been a
bit vague, but that a definition of a basic set of services (mostly
supervisors) would be defined and managed directly from this module.  In
further thinking, this was simply over-complicating things; in light of
the way we were developing MCP and the Platform, some additional
perspective has been cast on the issue.

The new idea is basically that we'll define a Behaviour
(`MscmpsSystInstance.Application`) which will define a basic API for
applications.  Each Application will be responsible defining its own
Application level services so long as there's a common interface for
starting and stopping instances of that Application.  What services,
supervisors, etc. constitute and Instance is entirely the responsibility
of the Application and its implementation.  So in this Component we're
really defining simple, common management across Applications, but that
we're really washing our hands for anything looking like an Application
or Instance implementation.  So long as the API defined in the behaviour
is satisfied... we're good.  Note that this simplifies and eliminates a
lot of service management code that we required when we were committed
to setting up a whole supervisory process for Applications and
Instances.

In addition to implementing that conceptual shift, we have the
implementation of NimbleOptions, the push for option definition and
validation to the API layer, string-named Datastores & Datastore
Contexts, and similar reformatting efforts as found elsewhere.