Refactor mscmp_syst_enums dev/test support
This changes the macros and service naming for dev and test service name
establishment.  Prior to this commit, we simply set both names such that
any consumer had to pick the correct dev or test name for its scenario.

In this commit, we change this process to resemble the same process in
mscmp_syst_db.  In that Component, we call the macro with a parameter
indicating either the `:dev` or `:test` nature of the context which turn
creates common variables with purpose specific values.

Update in all places where the reference is used.