Refactor enums and settings module function names
This refactor is targeted at simplifying and using common names across
modules.  While there is a certainty in the original, more specifically
named functions, it required more mental overhead to deal with and I'm
fairly sure to no reasonable benefit for the cost.

So simplify the naming of functions in the MscmpSystSettings &
MscmpSystEnums Components.  This mostly means removing the word
"setting" or "enum" where they exist directly in function names... the
module being used to call these functions should be apparent enough.

Because the APIs are changing, this means we need to regenerate the docs
for mscmp_syst_settings & mscmp_syst_enums as well.

Note that there are additional commits that will be related to this
change, but those commits will include the NimbleOptions related changes
and more of the larger-scale changes in this refactoring sequence.  This
commit is largely for those Components that have already been largely
refactored to include the additional changes here.