Refactor MscmpSystLimiter error tuples & unit test
This commit accomplishes two goals:

  - Push the MscmpSystError based error tuple generation to the public
    API layer consistent with the decision that such result tuples are
    a public API layer concern.

  - Fix some issues with the testing strategy including adding correctly
    structured unit tests.

Note that pushing the result tuple generation to the public API is a bit
different in this component as compared to others since we're just
really wrapping the hammer library which produces its own result tuples.
Still, we want to make those result tuples consistent with the rest of
our components... so it's not just a pass-through.