Refactor MscmpSystError re-imagining
This commit represents a significant refactoring of the MscmpSystError
Component.  The original Component was the first (or almost the first)
Component written and the first Elixir program of consequence I wrote.
In the time since that was original written, I have learned much about
Elixir development and structure as well as the program I'm writing.
However, this Component always felt weak.  Ideas here were brought from
a QtScript framework that I developed and which never quite fit here.
Also, early ideas about how this might be implemented were a touch
naive.  Finally, areas like the "code" field of the standard exception
never found a good, informative set of valid values and so ended up just
being set to `:undefined_error`, putting off setting real values to
tomorrow.

In this refactor, many of the aforementioned issues are addressed.  And
while I don't think it's yet arrived, my instinct suggests we're closer
now than when this Component was originally written.

The principle goals with this refactor include:

  - Create a framework where the other Components which depend on this
    Component can implement their own specific exception types while
    still conforming to a standard error handling process.

  - Make it more clear what errors are being produced and from where
    they originate.

  - Provide clarity for the definition of extra data that might be
    helpful in understanding the error (%MscmpSystError.Types.Context{})

This commit includes both the code for the MscmpSystError Component
itself and its documentation.

Finally, we bring the code more inline with the testing strategy that
developed after the original version of this Component was written.