Refactor error kinds to remove redundant naming
In MscmpSystUtilsData, MscmpSystNetwork, and MscmpSystLimiter, the new
exceptions local to each Component used "kind" names which had implicit
redundancies which didn't really add anything as far as understanding or
clarity.  For example `:parse_error` kind could easily be simple
`:parse` since this always appears in the context of
`Mserror.NetworkError`... there's no ambiguity or meaning that matters
outside of the immediate error generated or raised.

So remove those redundancies in the Components we've already touched and
regen the docs.