Refactor MscmpSystLimiter; result tuple to impl
This rectifies an issue I've been struggling with since I first decided
to deal with the error handling weaknesses.  Initially in the error
handling refactor, I was going to push as many result tuple decisions as
possible to the Public API layer assuming that with our unusual
bifurcation of the implementation and the API layers as being in
separate modules, we'd have better alignment of interests since
typically returning a result tuple is an API concern.

However, this actually runs afoul of the common practice and there are
two issues in particular that can come up constructing the result tuples
late in the API vs. early in the implementation/runtime logic.

  1. The implementation/runtime error reporting would be limited to
     raising with the exceptions being caught at the API for result
     tuple construction.  This encourages blanket rescues that can
     result in some exceptions being caught in result tuples which
     should raise since not all actions should generate a result tuple.

  2. The implementation/runtime logic returning result tuples allows for
     greater composability since you can include those calls in `with`
     statements and have sensible, controlled results.  This doesn't
     seem that compelling to be fair... most composition will be done by
     callers of the API and they get the result tuples in both the old
     and new methodologies.

But the result tuple logic in the implementation also means there's less
handling code at the API layer which is good, too since the API is less
about accomplishing real work than it is defining how users should
interact with the Component.  There will still be some code in the API
layer, chiefly instead of constructing result tuples, it will be focused
on deconstructing them where a raising version of an API function is
required; importantly, that shouldn't be in the implementation: choosing
to raise or accept a result tuple is an API concern and so the decision
should be there.... but it's less code at the API layer so that's
another reason to take that approach.