Refactor Components for refined exception handling
This impacts MscmpSystNetwork & MscmpSystLimiter, though this mostly
hits MscmpSystNetwork.

After working a bit with the network library I think I over-compensated
by pushing too much formalized error handling into the implementation
logic.  This refinement incorporates what feels like the final
structural opinion on the matter (typified by the MscmpSystNetwork
implementations here).

We do want greater use of result tuples in the implementation code, but
rather than these being formal MscmpSystError implementing exceptions
and errors, these can be simple, lightweight, and ad hoc.  The formal
MscmpSystErrors are only constructed and reported from the Public API
layer, incorporating the lower level errors as their "cause".

Ultimately we want the lower level to include enough information to
tell a client programmer what was found wrong and a location where the
problem was found; any other information may be included as needed, but
the basic form of a low level error will be `{:error, :some_reason}`
where the reason is unique enough to pinpoint where in the code the i
issue was identified, but also informative enough as to what the nature
of the problem was.  In some ways this can be seen as a "sub-error
message".  In general lower level errors will not be nested, but hoisted
as soon as possible to the API/caller.

At the API level, we return a formal MscmpSystError based error which
contains the complete messaging and what API function call failed.  As
Components get stacked these may become nested as lower level Component
calls bobble up to a higher level Component call.  Here the kind should
reflect something close to what the documented sub-divisions of the code
are... if the documentation is split into "parse", "management",
"services", the kind value of the error message should reflect that
(allowing the "cause" to be more specific about the exact nature of the
error).  Formal MscmpSystErrors existing only at the API level makes
them the tool of API communication when things are failing while
allowing lower levels to use less ceremony to achieve similarly useful
goals in regard to error management.

There are a fair number of changes prompted by this update including
some code & test cleanup and a documentation regeneration.