Add MscmpSystUtilsGuards Component
This Components provides guard definitions that are generally useful and
which don't clearly belong to any other specific component.

Currently only one guard is included (is_reg_atom/1).  It's hard to tell
if this is really enough to justify a whole Component for guards, but we
can reevaluate that later.

This commit includes the documentation as well, for what it's worth.