Add devcontainer setup
Trying to use devcontainers to avoid the setup churn that directly
developing on a local environment brings.

Right now I'd call this a mixed experience.  On the one hand, the basic
idea is great: I can pull the elixir image, the postgres image, and have
a repeatable setup.  However the process is quirky, doesn't "just work",
and requires a fair amount of futzing to get right.  My sense is that
this is really trying to use a technology (Docker Containers), which are
meant to run applications on servers, as an interactive development
technology... which it just isn't really meant for.

Anyway, we'll continue for now.  Note that my setup is still a work in
progress and probably not very good as it stands.