Add devcontainer setup
This commit of the devcontainer setup is only being made for historical
purposes.  While I had this working with VS Code and its related
extensions, I had it going on WSL2 and file operations performing so
poorly were simply not tenable; apparently a well known issue with no
fixes or worthwhile workarounds.

If this comes back up we'll at least have pertinent examples to draw
from.