Refactor MscmpSystDb for Mserror.DbError
This updates this component for the changes recently made to
MscmpSystError.

To be fair the code in MscmpSystDb isn't great and this commit doesn't
change this.  It's clear that we need to be sure that our specialized
connections (DBA & Privileged) are correctly terminated without regard
to a extenuating error condition stopping a transaction early.  Some of
this has been closed here, but prior to any release, this would be a
housekeeping item.

Also. it's pretty clear the tests are crap.  These really need to be
made more comprehensive and in conformity with our broader testing
strategies; to be fair, this was an early module before much of that had
been thought out (and before I had much Elixir experience), but it's
clear now that this area also needs improvement.