Refactor specialized database reset scripts
In original implementation, if you were doing mscmp_syst_db or
mscmp_syst_instance dev/testing, you'd have to run the specialized
database cleanup script defined by the Component which, in turn, would
call the generalized database clean up script.  This worked, but meant
the developer had to remember special stuff depending on the Component
being worked with.

This change resolves that by having the general clean up script call the
specialized cleanup scripts whether it needs to or not.  This way,
there's only a single database clean up process that applies in all
circumstance.  The only special thing needed to remember is if the
special clean up scripts themselves need to be changed, then they need
to be changed in the Component files itself.  This is actually OK,
because we still want the Component to be as defined in a single place
as much as possible... we want to minimize specific Component references
outside dependency declaration as much as possible.

Note that as this commit implements things, it may not work as-is.  I've
tested in the "nothing to do" circumstance and that works well enough;
but the specialized scripts both did some early specialized work, called
the standard scripts, and then finished the specialized work.  I'm not
sure if this phased approach is required because of the old way of doing
things or if the new way needs to duplicate that split: I'm hopeful that
is won't need to be done since we're being more true to the object
creation order in this commit.