Refactor MscmpSystAuthn for MscmpSystError upd
Fairly extensive review and update of the component in light of the
changes made to MscmpSystError.

Note that in this commit Dialyzer was reporting two pattern can never
match errors that have had attributes added to ignore.  I've spent some
solid time trying to understand what Dialyzer doesn't like about these
two functions, both of which are in the Public API module MscmpSystAuthn
and I just can't find it; tests are passing and other, very similar
functions are passing and don't appear to have any substantial
differences.  For now then, give up/move on and trust in the tests.  It
bothers me that these exceptions are there, so I do think it should be
revisited, but the issue simply isn't warranting additional time to be
spent on resolution at this point.