Add get_pg_exception to MscmpSystDb
This function adds Postgrex.Error exception parsing, primarily to reduce
these exceptions to simple `{:kind, "message"}` style tuples for later
return in error tuples.  This is especially needed since we have
database logic which can raise PostgreSQL exceptions for transaction
terminating, but non-fatal overall, validation issues.  These custom
SQLSTATE values from PostgreSQL get returned in the
`%Postgrex.Error{postgres: %{pg_code: <SQLSTATE>}}` exception struct on
the Elixir side.  It's this that we parse in this change to make it more
"Elixiry" on the application code side.

Note that the Postgrex.Error structure, according to community
conventions is likely taking advantage of an internal implementation
detail of Postgrex that may not be supported over time or may see
unannounced breaking changes.