Add error result parsing to MscmpSystError
After the original refactoring to Mserror based exceptions and using
result tuples more broadly for internal logic, I found that while many
matters improved, reporting errors because a bit of a chore.  The API
level errors coming out of a Component I wanted to be fairly clean, with
the only nesting really being from lower level, nested API calls.  This
meant I was destructuring error tuples and other exceptions to get to
that clean picture every time I wanted to report an error.  Also, I was
a bit worried that, in cleaning things up, I couldn't always communicate
a clear message from the lower level since the only thing that really
survived at the conversion to a Mserror struct was the `cause`, which
was intended to be terse and often easy to match against.

The solution to both issues is the implementation of a new `parse_error`
option to the `new/3` constructor for Mserror based exceptions.  This
allows pretty much any arbitrary error tuple to be passed to the the
constructor, allowing the cause and the message to both be set from the
lower level error tuple, so long as it is in the form:
`{:error, {:cause, "message"}}`.  Other forms of error tuples or even
any other term can be sent as the `parse_error` option, though what can
be derived will vary based on the specific form provided.

This should be backward compatible with the other implementation so I
don't think we necessarily have to do another big error related
refactoring right now, thought we can improve things as we go.