Refactor MscmpSystPerms for most recent error upd
This brings the less critical improvements for error and database
exception handling over to MscmpSystPerms. Mostly because I want a more
intensive, but still light, test of the last changes to error handling.

I think at this point it looks good and works well.  With Cursor/AI
helping with the refactoring it didn't really take that long.  Still,
no rush to push this most recent change out even though I think it helps
the simplicity cause, consistency, and overall pleasantness a fair
amount.  We'll do it piecemeal and as I encounter (and am annoyed by)
it.