Refactor MscmpSystUtils & deps version bump
The primary change in this commit is taking the previously started
MscmpSystUtilsProcess component and instead consolidating that and
MscmpSystUtilsString and MscmpSystUtilsGuards into a single
MscmpSystUtils component.

As we find other areas where we need small utilities here and there, it
makes much more sense to manage these in a single Elixir project than to
start a new project for every possible kind of utility that we might
need.  So we consolidate all "basic" utilities into a single Elixir
project while retaining the naming.

This consolidation does not include the MscmpSystUtilsData Component.
That component brings in Ecto and all of its supporting dependencies
which, if the data piece is not needed seems wasteful.  So we'll
continue separating our heavy-weight utilities into their own Elixir
projects while allowing the low/no dependency utilities to live in the
MscmpSystUtils Component; this dependencies difference is what we mean
by "basic".

Other changes in this commit include getting rid of the
`use Msutils.Guards` option and related macros preferring to just
import Msutils.Guards.... which was all the macro was doing anyway.
Change all dependents from use to import as well.

Finally, we bump overall project dependencies in hope that some have
been updated to avoid Elixir 1.18 deprecation and typing errors.