Refactor MscmpSystInstance type names
These are just more direct in describing these types in terms of
categories rather than something that is more descriptive of a specific
parameter use case.

Also change the exception type required by the reported result.
Previously the MscmpSystInstance Component itself was doing a fair
amount of Instance heavy lifting, but that's now all on the implementing
subsystem and so needs to be more generalized in MscmpSystInstance.