Add MscmpSystUtils.Process registration support
If we're dealing with GenServers or Supervisors directly, Elixir has
existing naming and name/registry conventions which are completely
sufficient.  However, in a number of cases we don't have directly access
to those mechanisms; for example, using a process registry for Ecto.Repo
is currently a non-starter with the library itself.  Same is true for
PubSub and possibly other process which can be started and messaged as
long-lived processes.

The goal with this change is not to supersede those cases where the
process naming convention is completely workable (anything which accepts
`GenServer.name/0` style names), but to have a common way of dealing
with those services or our own services where that's not currently an
option.

In this change we add the ability for our own services to register names
using the `GenServer.name/0` style (mostly).  Further we extend use this
capability throughout our stack as appropriate (such as in test
Registries).  It won't be used everywhere we have process naming, but
fills a gap that should have us completely covered in this regard.

An interesting aside is that there a GenServer function which can parse
values conforming to `GenServer.name/0`, but it's marked private and
so we don't want to just depend on that.... of course, we could have
wrapped that function in our API to guard against sudden GenServer
changes forcing large changes in our reliance on that code... but I'm
not so smart and realized that only after I was done implementing our
version of those functions.