Add ms_syst_documentation database role
This adds the `ms_syst_documentation` role as a standard database role
in the bootstrap process.  The role is a non-login, low-privilege role
which can enumerate objects and read their definitions from the catalog
but not the data.

The intention is that any named user generating documentation will add
their personal database role as a member of the `ms_syst_documentation`
role.

Finally note that the database clean-up scripts such as the reset script
do not remove this account; doing so would force the named users to have
to rejoin the role which would be unnecessarily cumbersome.  I think so
long as this behavior is reasonably documented we'll be OK.