Remove older msplatform
There's been sufficient churn and ideas around the structure of the
project, and little work done on the Platform, that we can basically
start it from scratch.  We've got a copy of this code saved off so we
aren't actually going to absolutely reinvent the wheel, but enough so
that we can treat msplatform as a new start when we get there.  Finally,
Phoenix itself has gone through a fair amount of change so starting with
a fresh Phoenix project when we're ready will make more sense than also
trying to update the older Phoenix stuff.

This commit removes the old msplatform so that, when we're ready, we can
start it back up on newer Phoenix versions.