Refactor build tool from Powershell to Elixir
While Powershell was a good choice for our scripts, it became apparent
that we could be more consistent with the project to actually use Elixir
in a scripting context to manage our build scripts.  I was writing
Powershell infrequently enough that either I was completely reliant on
the LLM to be productive at all or getting things done very slowly...
keeping the scripts in Elixir meant that I could more readily jump in
and get things done as needed.

  - Migrate existing functionality from Powershell to Elixir

  - Add database cleaning functionality

  - Add test running functionality

  - Add some migration management functionality

All of these changes should make the tool more useful and, importantly,
to use the tool as the primary test runner in CI contexts.