Refactor MscmpSystDb drop_datastore functionality
This addresses a few issues when handling the dropping of datastores in
some contexts.  An example of a special handling scenario is the
generation of database documentation.  We use the normal migration
loading process to create the database and load the schema, but we
don't ever actually start Elixir DB connections (or even an Elixir app
for that matter) and drop_database wanted to be sure that we stop such
services... leading to an error.

The primary addition to deal with this sort of scenario is the addition
of an option to bypass dealing with Datastore Context connections.

Along the way we also address some Context Registry needs in terms of
identifying a registry module and registry name.  This allows higher
level Components and Subsystems to define their own registry type and
registry so long as it conforms to the norms.