Refactor ms_syst_docuemtnation role drop on clean
Previously we didn't drop the ms_syst_documentation role on cleaning the
database because we needed to assign arbitrary, personal roles to the
group role.

Now the build tools script automatically assigns the current database
user (the one used to run the reset script) to the ms_syst_documentation
role and it's most common that no other roles will be members... so we
go ahead and favor a fuller reset vs. leaving objects behind from the
clean process.