Remove old mssub_mcp files
We're in the process of re-writing this Component with some new thinking
about how things work.  Rather than try to show an evolution from old to
new, it makes more sense to simply discard the old outright since that's
the actual process we're working through.

It's not that some of this older stuff is gone for good, I've kept a
copy of the old code and am referencing/reusing frequently when it
makes sense to do so... but I did literally move the mssub_mcp directory
out of the source tree and recreated it from scratch.