Refactor MscmpSystDb testing; align with strategy
This just makes the MscmpSystDb testing strategy aligned with that found
in other Components.  The reality is that it was mostly there already.

Now this isn't to say that this makes the testing sufficient... it's
still lackluster... but it is no worse and is now on a better foundation
which can be expanded on (corrected?) in future.