Fix MsbmsBuildLib/TestsElixir handling test fails
The previous code couldn't handle multiple test failures because certain
kinds of test failures leave the testing database objects in place; this
in turn meant that we needed to explicitly and out-of-band clean the
database, but we don't have the needed info at this point to do that.

We could resolve the issue, but getting the required parameters this
deep into the test scripts makes less sense than just halting the test
progress early.

Therefore in this commit we simplify the TestsElixir to not try to run
all tests requested and collecting results as we go; rather we simply
halt progress at the first testing failure condition.