Fix MscmpSystInstance bad registry errors
There are a couple cases where we call MscmpSystDb.drop_datastore/2
which in turn caused the registry error to show up because we don't
provide a registry in these contexts.  A registry is required because
drop_datastore now wants to stop any related contexts prior to dropping
the database.

In the cases addressed in this commit, we shouldn't have running
contexts.  The first case is purging Instances.  Purge eligible
instances shouldn't have running contexts which need stopping.  The
other case is dropping a newly created datastore which failed
initialization.  Again, an uninitialized datastore shouldn't have
running contexts that need stopping.

In both cases, simply bypass the stopping phase of drop_datastore/2.