Refactor msbms_build.exs default log level
Info is the more sensible default for streaming messages in the console.
Console usage is the much more common case, so cater to that.

If CI is running, we can always use the command line option to tone down
the logging to the essential (i.e. "notice", or even "warning").