Migrate msbms_build.exs to MsbmsBuildLib Mix Tasks
The script was a fine idea, but why not do what all of these mix
generators do and define our code maintenance tasks as mix tasks?

This commit migrates the now removed msbms_build.exs script
functionality into similar mix tasks.

Add the ability to define component lists as files rather than requiring
them to always be set at the command line.  This should be useful for
having planned CI runs.

Add documentation and verify typespecs where required.

Note that this was largely an AI coded effort.  Naturally I edited and
guided, but I didn't nitpick and this isn't regular production code...
so the balance between fast and good favors fast.  At this commit the
documentation is OK, but could be improved explaining the context of the
build library and how to build and deploy it as an archive.