Refactor Elixir component mix.exs dependency mngmt
Prior to this the mix.exs of each project was very conventional in the
way it defined versions of dependencies, Elixir version, etc.; ignoring
some minor structural things inside the file (i.e. more module attribute
defined values than typical).  While this works, it causes us to have to
visit every Component to update them when dependency versions change or
our Elixir version changes, etc.  But we haven't had a need to have
different versions of dependencies, etc. across the project.  In fact,
it's a goal to have common dependencies across the project to be exactly
the same.

To this end, and since mix.exs does its thing as a runnable Elixir
script, we are making a common build time configuration which can be
eval'ed by all the mix.exs scripts.  We just say what dependencies we
want and the global config file loads the correct versions into the
running mix.exs.  All version maintenance is therefore in the config
file.