Add project scaffolding to MsbmsBuildLib
Our standard Elixir Component is not the standard project layout and so
`mix new` leaves a lot to do manually each time we start a new
component.

This new component scaffolding follows our project convention.

Note that this, again, is largely an AI effort which review.  Being non-
production code, I've not intervened or contradicted as much as I might
otherwise as time in more valuable than achieving perfections.